requirements.txt
aiohttp==3.5.4
APScheduler==3.6.1
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
discord.py==1.2.3
idna==2.8
lxml==4.4.0
multidict==4.5.2
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
tzlocal==2.0.0
urllib3==1.25.3
websockets==8.0.2
yarl==1.3.0
